Style tweak - dropped form from synergy view
Biggest change was removing the submission form from the view page,
since it was taking up lots of space and encourages an iteration process
which hurts the search feature.  This doesn't mean that the problem is
solved - there needs to be a good way to quickly change things while
still retaining the ability to filter "partial submissions" from the
search list.

Removed some of the row-fluid classes, since a simple container did the
job just fine.  Also removed some of the internal containers on view -
flat design feels better and <hr> does the job just fine.