Large rename synergy name -> title
First thing I noticed with the new management tool was dozens of people
putting a card name in the name field  - I can see where this would be
unclear.

I have some other ideas to make it clearer where things go, but for now
this should help out a bit.